+++ doc/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - XML code completion was improved: entries are sorted alphabetically, attribute suggestions are provided based on
   the current entity.
 - JSON Code Completion support was added in particular for PKS Edit configuration files.
+- The new function `reload-themes` (by default bound to `Ctrl+F10`) makes theme updates and editing easier.
 
 ### Fixed
 - A potential crash was fixed, which occurred, when opening a dialog displaying the *currently selected text* (e.g.